requirements.txt
numpy
scipy
torch
google-generativeai
typing-extensions
Pillow
transformers
huggingface-hub
diffusers
python-dotenv
requests 
rembg
onnxruntime
